# openapi.yaml
openapi: 3.0.3
info:
  title: Woliz Products API
  version: 2.0.0
  description: >
    API for managing the Woliz product catalog. The catalog uses a temporal
    model: a product keeps a stable identity, but its characteristics
    (price, format, packaging, SKU) evolve over time via variants.
    Products can belong to the global catalog (nanostoreId = null) or to a
    specific nanostore (private product). Includes user registration,
    login, and JWT-based authentication.
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
paths:
  /products:
    get:
      summary: List all products
      operationId: listProducts
      security: []
      parameters:
        - name: nanostoreId
          in: query
          description: >
            Filter by nanostore. Omit or leave empty to list the global
            catalog (products with nanostoreId = null).
          required: false
          schema:
            type: integer
        - name: locale
          in: query
          description: Locale used to resolve product translations in the response
          required: false
          schema:
            $ref: '#/components/schemas/Locale'
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      summary: Create a new product
      description: >
        Creates a product along with its required translations (at least
        one in ar-ma) and its first variant. Writes a sku_history entry
        with changeType = creation.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input (e.g. missing ar-ma translation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
        '409':
          description: SKU or barcode conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/bulk:
    post:
      summary: Push multiple products at once
      operationId: createProductsBulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 500
              items:
                $ref: '#/components/schemas/NewProduct'
      responses:
        '201':
          description: Products created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
  /products/{id}:
    get:
      summary: Get a product by id
      operationId: getProduct
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1000
        - name: locale
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Locale'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update an existing product's stable fields
      description: >
        Updates fields on the products table itself (brand, category,
        image, status...). To change price, size or packaging, use the
        variants endpoint instead, since those changes create a new
        variant rather than overwriting the existing one.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /partenaires:
    get:
      summary: List all partners
      operationId: listPartenaires
      security: []
      responses:
        '200':
          description: A list of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Partenaire'

    post:
      summary: Create a new partner
      operationId: createPartenaire
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPartenaire'
      responses:
        '201':
          description: Partner created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partenaire'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
  /partenaires/{partenaireId}/villes:
    get:
      summary: List cities for a partner
      operationId: listVilles
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ville'
        '404':
          description: Partner not found

    post:
      summary: Add a city to a partner
      operationId: createVille
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVille'
      responses:
        '201':
          description: City created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ville'
        '400':
          description: Invalid input
        '401':
          description: Missing or invalid token
        '404':
          description: Partner not found
...
  /partenaires/{partenaireId}/villes/{villeId}:
    get:
      summary: Get a city by id
      operationId: getVille
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: City found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ville'
        '404':
          description: Partner or city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a city
      operationId: updateVille
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVille'
      responses:
        '200':
          description: City updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ville'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Partner or city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a city
      operationId: deleteVille
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: City deleted
        '404':
          description: Partner or city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /partenaires/{partenaireId}/villes/{villeId}/toggle:
    patch:
      summary: Toggle a city's active status
      operationId: toggleVille
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: City toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ville'
        '404':
          description: Partner or city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
  /partenaires/{partenaireId}:
    get:
      summary: Get a partner by id
      operationId: getPartenaire
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Partner found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partenaire'
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a partner
      operationId: updatePartenaire
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPartenaire'
      responses:
        '200':
          description: Partner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partenaire'
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a partner
      operationId: deletePartenaire
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Partner deleted
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /partenaires/{partenaireId}/toggle:
    patch:
      summary: Toggle a partner's active status
      operationId: togglePartenaire
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Partner toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partenaire'
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Locale:
      type: string
      enum: [en, fr, ar, ar-ma]
    CreatedByType:
      type: string
      enum: [admin, user_brand]
    VariantStatus:
      type: string
      enum: [active, scheduled, expired, archived]
...
    NewProduct:
      type: object
      additionalProperties: false
      required:
        - sku
        - categoryId
        - price
        - translations
        - initialVariant
      properties:
        sku:
          type: string
          maxLength: 50
          example: "WOL-1000"
        barcode:
          type: string
          nullable: true
          example: "6111234567890"
        unit:
          type: integer
          nullable: true
          description: Units per pack, e.g. 6 for a pack of 6
          example: 6
        price:
          type: number
          format: float
          example: 19.99
        brandId:
          type: integer
          nullable: true
          description: Nullable since June 2026 for products without a brand
        categoryId:
          type: integer
          description: Required
        type:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
          description: Legacy R2 path field
        imageMediaId:
          type: integer
          nullable: true
        images:
          type: array
          items:
            type: string
          nullable: true
        stock:
          type: integer
          default: 0
        minStock:
          type: integer
          default: 0
        status:
          type: boolean
          default: true
        nanostoreId:
          type: integer
          nullable: true
          description: >
            Null = global catalog product managed by Woliz, visible to all
            nanostores. Set = private product owned by one nanostore.
        zoneId:
          type: integer
          nullable: true
          description: >
            Zone the product is sold in. Determines the product's city and
            partner via the zone's own ville/partenaire relations.
        translations:
          type: array
          minItems: 1
          description: At least one translation in ar-ma is required.
          items:
            $ref: '#/components/schemas/NewProductTranslation'
        initialVariant:
          $ref: '#/components/schemas/NewProductVariant'
    UpdateProduct:
      type: object
      additionalProperties: false
      minProperties: 1
      description: >
        Only stable product fields can be updated here. Price, size and
        packaging changes must go through /products/{id}/variants.
      properties:
        brandId:
          type: integer
          nullable: true
        categoryId:
          type: integer
        type:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        imageMediaId:
          type: integer
          nullable: true
        images:
          type: array
          items:
            type: string
          nullable: true
        minStock:
          type: integer
        status:
          type: boolean
        zoneId:
          type: integer
          nullable: true
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1000
        sku:
          type: string
          example: "WOL-1000"
        barcode:
          type: string
          nullable: true
          example: "6111234567890"
        unit:
          type: integer
          nullable: true
        price:
          type: number
          format: float
        unitPrice:
          type: number
          format: float
          nullable: true
        brandId:
          type: integer
          nullable: true
        categoryId:
          type: integer
        type:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        imageMediaId:
          type: integer
          nullable: true
        images:
          type: array
          items:
            type: string
          nullable: true
        stock:
          type: integer
        minStock:
          type: integer
        status:
          type: boolean
        createdBy:
          type: integer
          nullable: true
        createdByType:
          $ref: '#/components/schemas/CreatedByType'
        nanostoreId:
          type: integer
          nullable: true
        zoneId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        translations:
          type: array
          items:
            $ref: '#/components/schemas/ProductTranslation'
        activeVariant:
          $ref: '#/components/schemas/ProductVariant'
        creator:
          $ref: '#/components/schemas/User'
        zone:
          $ref: '#/components/schemas/Zone'

    # ---- Product translations ----
    NewProductTranslation:
      type: object
      additionalProperties: false
      required:
        - locale
        - name
      properties:
        locale:
          $ref: '#/components/schemas/Locale'
        name:
          type: string
          maxLength: 255
          example: "Souris sans fil"
        description:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
          example: "500ml"
        packaging:
          type: string
          nullable: true
          example: "pack de 6"
    ProductTranslation:
      allOf:
        - $ref: '#/components/schemas/NewProductTranslation'
        - type: object
          properties:
            id:
              type: integer
            productId:
              type: integer

    # ---- Product variants ----
    NewProductVariant:
      type: object
      additionalProperties: false
      required:
        - sku
        - price
        - validFrom
      properties:
        sku:
          type: string
          maxLength: 100
          example: "WOL-1000-V1"
        price:
          type: number
          format: float
          example: 19.99
        costPrice:
          type: number
          format: float
          nullable: true
        size:
          type: string
          nullable: true
          example: "500ml"
        packaging:
          type: string
          nullable: true
          example: "bouteille"
        stock:
          type: integer
          default: 0
        minStock:
          type: integer
          nullable: true
        maxStock:
          type: integer
          nullable: true
        validFrom:
          type: string
          format: date
        validUntil:
          type: string
          format: date
          nullable: true
          description: Null means unlimited
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        translations:
          type: array
          items:
            $ref: '#/components/schemas/NewProductVariantTranslation'
    ProductVariant:
      allOf:
        - $ref: '#/components/schemas/NewProductVariant'
        - type: object
          properties:
            id:
              type: integer
              example: 5000
            productId:
              type: integer
            isActive:
              type: boolean
            status:
              $ref: '#/components/schemas/VariantStatus'
            createdBy:
              type: integer
              nullable: true
            updatedBy:
              type: integer
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # ---- Product variant translations ----
    NewProductVariantTranslation:
      type: object
      additionalProperties: false
      required:
        - locale
        - name
      properties:
        locale:
          $ref: '#/components/schemas/Locale'
        name:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        packagingLabel:
          type: string
          nullable: true
...
    NewPartenaire:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          example: "Woliz Distribution"
        contact:
          type: string
          nullable: true
          example: "Ahmed Benali"
        email:
          type: string
          format: email
          nullable: true
          example: "contact@woliz-distrib.ma"
        phone:
          type: string
          nullable: true
          example: "+212600000000"
        status:
          type: string
          enum: [actif, inactif]
          default: "actif"  
    Partenaire:
      allOf:
        - $ref: '#/components/schemas/NewPartenaire'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            villes:
              type: array
              items:
                $ref: '#/components/schemas/Ville'
    NewVille:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          example: "Rabat"
        status:
         type: string
         enum: [actif, inactif]
         default: "actif"    
    Ville:
      allOf:
        - $ref: '#/components/schemas/NewVille'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            partenaireId:
              type: integer
              example: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            zones:
              type: array
              items:
                $ref: '#/components/schemas/Zone'
    NewZone:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          example: "Agdal"
        endpoint:
          type: string
          nullable: true
          description: >
            Route/URL used to reach this zone's product feed, e.g.
            api/:fourniseur/:ville/:zone/product
          example: "api/woliz-distrib/rabat/agdal/product"
        status:
         type: string
         enum: [actif, inactif]
         default: "actif"  
    Zone:
      allOf:
        - $ref: '#/components/schemas/NewZone'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            villeId:
              type: integer
              example: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # ---- Users / auth ----
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane@example.com"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "email ou mot de passe incorrect"